+++ pkg/debrid/types/client.go
@@ -8,7 +8,7 @@ type Client interface {
 	SubmitMagnet(tr *Torrent) (*Torrent, error)
 	CheckStatus(tr *Torrent, isSymlink bool) (*Torrent, error)
 	GenerateDownloadLinks(tr *Torrent) error
-	GetDownloadLink(tr *Torrent, file *File) (string, string, error)
+	GetDownloadLink(tr *Torrent, file *File) (*DownloadLink, error)
 	DeleteTorrent(torrentId string) error
 	IsAvailable(infohashes []string) map[string]bool
 	GetCheckCached() bool
@@ -18,9 +18,10 @@ type Client interface {
 	GetName() string
 	GetLogger() zerolog.Logger
 	GetDownloadingStatus() []string
-	GetDownloads() (map[string]DownloadLinks, error)
+	GetDownloads() (map[string]DownloadLink, error)
 	CheckLink(link string) error
 	GetMountPath() string
 	DisableAccount(string)
 	ResetActiveDownloadKeys()
+	DeleteDownloadLink(linkId string) error
 }